chore: rename src/local-invoke/ → src/local/ ahead of cdkd local start-api (#224)
Independent small PR ahead of `cdkd local start-api` (PR 5 of #224).
The new directory will host both `cdkd local invoke`'s building blocks
AND the upcoming `start-api` server / container-pool modules, so the
file-system shape mirrors the `cdkd local *` CLI family.

Pure rename:

- `git mv src/local-invoke src/local`
- `git mv tests/unit/local-invoke tests/unit/local`
- Every `from '../local-invoke/...'` import rewritten to `from '../local/...'`

Integration test directory names (`tests/integration/local-invoke{,-python,-from-state}/`)
stay unchanged — those are user-facing test names referenced by `/run-integ`
and renaming them would break that surface unrelated to the source-path
concern.

CLAUDE.md: directory description at the top of "Important Files" updated
to point at the new path; PR 2's clickable file links updated; new
"Recently Implemented" bullet added. PR 1 / PR 2 / PR 4's existing
prose references to `src/local-invoke/...` are left as historical record
(per CLAUDE.md's own "do not modify existing bullets" rule).

No behavior change.
