+++ src/local/lambda-resolver.ts
