+++ src/local/runtime-image.ts
